Centralize analysis flag derivation and fix "is in focus" checks
