+++ src/ui/actions/logpoint.ts
@@ -254,7 +254,7 @@ async function setMultiSourceLogpoint(
         showPrimitiveLogpoints(logGroupId, points.data || [], primitiveFronts);
       }
       // If we're only displaying only primitives, we can bail out if there's no condition or the condition request is done
-      if (!condition) {
+      if (!condition && points.isCompleted) {
         return;
       }
     }
